Split future code into separate module
Better to not have to worry about future generalizations and assess the code for them at the moment
