+++ setup.cfg
@@ -52,12 +52,12 @@ install_requires =
     y0>=0.2.5
     scipy
     numpy
-    ananke-causal
-    pgmpy
+    ananke-causal>=0.5.0
+    pgmpy>=0.1.24
     matplotlib
     pandas
     seaborn
-    optimaladj
+    optimaladj>=0.0.4
 
 # Random options
 zip_safe = false
